+++ yggdrasil-web/src/api/mod.rs
@@ -1,5 +1,6 @@
 pub mod admin;
 pub mod me;
+pub mod profiles;
 pub mod scores;
 pub mod universes;
 pub mod user_profiles;